Add macOS 10.11.x support

+++ cdi4dc/inc/cdibuild.h
@@ -20,6 +20,6 @@
 int write_gap_tracks(FILE* cdi);
 int write_data_track(FILE* cdi, FILE* iso);
 void write_cdi_header_start(FILE* cdi, char* cdiname);
-void write_cdi_head_end(FILE* cdi, char* volumename, long total_cdi_space_used, long cdi_end_image_tracks);
+void write_cdi_head_end(FILE* cdi, char* volumename, uint32_t total_cdi_space_used, long cdi_end_image_tracks);
 
 #endif // __CDIBUILD__H__
